+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-* Updated translations.
+* Added option "Auto add new mounts to selected" in profile settings.